requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
fpdf
python-docx
kaleido
